+++ lambda-assets/verifiers-fetcher/app.ts
