+++ src/index.ts
@@ -1,5 +1,5 @@
 export { JustInTimeRegistration } from './just-in-time-registration';
-export { VerifiersFetcher } from './verifiers-recorder';
+export { VerifiersFetcher } from './verifiers-fetcher';
 export { CaRegistrator } from './ca-registrator';
 export { DeviceActivator } from './device-activator';
 export { ReviewReceptor, ReviewAcceptionRole } from './review-receptor';
